requirements.txt
Flask
flask-cors
opencv-python-headless
python-doctr[torch]
fpdf
